+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Piral Changelog
 
+## 1.1.0 (tbd)
+
+- Fixed retrieval of dep versions not exporting their *package.json*
+- Fixed calling custom Webpack config throwing an exception
+- Fixed absolute to a relative path for shared dependencies
+- Updated dependencies of `piral-cli-webpack5`
+- Added support for optionally centrally shared dependencies
+- Added possibility to extend notification types to `piral-notifications`
+- Added possibility to further extend options w.r.t. menu type in `piral-menu`
+- Added lint rule for detecting potential CSS conflicts in pilets (#611)
+
 ## 1.0.2 (June 27, 2023)
 
 - Fixed support for `piral-extension` inside shadow DOM